Split bot & worker into separate programs
