+++ README.md
@@ -6,10 +6,11 @@ A discord bot, written using [discordgo](https://github.com/bwmarrin/discordgo),
 
 - Build all dependencies with custom permitted cflags, to build the ImageMagick wrapper:
   ```shell
-  CGO_CFLAGS_ALLOW=-Xpreprocessor go build -a
+  CGO_CFLAGS_ALLOW=-Xpreprocessor go build -a ./cmd/borik ./cmd/borik-worker
   ```
 - Copy `.env.dist` to `.env`, and populate it with a token and a prefix
-- `go run . run`
+- Run the bot with `go run ./cmd/borik`
+- Run the worker with `go run ./cmd/borik-worker`
 
 ### Nix
 
@@ -19,7 +20,7 @@ Start by following the `.env` instructions above, then do one of the following:
 
 #### Running for usage
 
-If you just want to use Borik and don't intend on working on it yourself, running `nix run` should be all that is required to compile and start Borik.
+If you just want to use Borik and don't intend on working on it yourself, running `nix run .#bot` and `nix run .#worker` should be all that is required to compile and start each component.
 
 #### For development
 
@@ -27,4 +28,4 @@ If you plan on working on Borik, run the below commands to prepare a dev shell a
 
 - Run `nix develop`
   - This will drop you into a shell with Go & all required dependencies ready to go.
-- Run `go run .`
+- Run `go run ./cmd/borik` or `go run ./cmd/borik-worker`
