+++ cmd/borik/main.go
@@ -1,6 +1,7 @@
-package cmd
+package main
 
 import (
+	"fmt"
 	"os"
 	"os/signal"
 	"syscall"
@@ -9,11 +10,12 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/fogo-sh/borik/pkg/bot"
+	"github.com/fogo-sh/borik/pkg/config"
 )
 
-var runBotCmd = &cobra.Command{
-	Use:   "bot",
-	Short: "Run the Discord bot",
+var rootCmd = &cobra.Command{
+	Use:   "borik",
+	Short: "Run the Borik Discord bot",
 	Args:  cobra.NoArgs,
 	Run: func(cmd *cobra.Command, args []string) {
 		borik, err := bot.New()
@@ -28,16 +30,30 @@ var runBotCmd = &cobra.Command{
 			}
 		}()
 
-		log.Info().Msg("Borik is now running, press CTRL-C to exit.")
+		log.Info().Msg("Borik bot is now running, press CTRL-C to exit.")
 		sc := make(chan os.Signal, 1)
 		signal.Notify(sc, syscall.SIGINT, syscall.SIGTERM, os.Interrupt)
 		<-sc
-		log.Info().Msg("Quitting Borik")
+		log.Info().Msg("Quitting Borik bot")
 
 		borik.Stop()
 	},
 }
 
 func init() {
-	runCmd.AddCommand(runBotCmd)
+	cobra.OnInitialize(loadConfig)
+}
+
+func loadConfig() {
+	err := config.Load()
+	if err != nil {
+		log.Fatal().Err(err).Msg("Error loading config")
+	}
+}
+
+func main() {
+	if err := rootCmd.Execute(); err != nil {
+		_, _ = fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
 }
