--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/fogo-sh/borik/cmd"
-
-func main() {
-	cmd.Execute()
-}
